wifi:Stub out support for SET_AP_WPS_P2P_IE command
The supplicant is sending SET_AP_WPS_P2P_IE driver command
which is not handled by driver. This is an API that is directly
called by Android extensions to the wpa_supplicant. As the
SET_AP_WPS_P2P_IE command is not handled, the handler function
for this command is removed in wpa_supplicant_8_lib.

Change-Id: Ic3f55c775d176edd2472414eefc447b9da084048
CRs-Fixed: 821067